Remove cmake based setup, build CI using setup script.
Do not fetch the main project folder in setup script, it messes up the
drone setup as the branch created there does not actually track the
origin/master branch. This is a general issue: if the HEAD is not
tracking anything, the setup script cannot be used in its current form.